+++ Assets/UGF.Testing.Runtime.Tests/Scopes/TestTestGameObjectScope.cs
@@ -1,8 +1,8 @@
 using NUnit.Framework;
-using UGF.Testing.Runtime.Tests.Scopes;
+using UGF.Testing.Runtime.Scopes;
 using UnityEngine;
 
-namespace UGF.Testing.Runtime.Tests.Tests.Scopes
+namespace UGF.Testing.Runtime.Tests.Scopes
 {
     public class TestTestGameObjectScope
     {
