+++ Assets/UGF.Testing.Runtime.Tests/Scopes/TestTestLogEnableScope.cs
@@ -1,10 +1,10 @@
 using NUnit.Framework;
-using UGF.Testing.Runtime.Tests.Scopes;
+using UGF.Testing.Runtime.Scopes;
 using UnityEngine;
 
-namespace UGF.Testing.Runtime.Tests.Tests.Scopes
+namespace UGF.Testing.Runtime.Tests.Scopes
 {
-    public class TestTestLogScope
+    public class TestTestLogEnableScope
     {
         private bool m_logReceived;
