+++ Packages/UGF.Testing/Runtime/Scopes/TestComponentScope.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UGF.Testing.Runtime.Tests.Scopes
+namespace UGF.Testing.Runtime.Scopes
 {
     /// <summary>
     /// Represents a gameObject with component scope.
