+++ Packages/UGF.Testing/Runtime/Scopes/TestObjectScope.cs.meta
