+++ Packages/UGF.Testing/Runtime/UGF.Testing.Runtime.asmdef
@@ -1,6 +1,6 @@
 {
-    "name": "UGF.Testing.Runtime.Tests",
-    "rootNamespace": "",
+    "name": "UGF.Testing.Runtime",
+    "rootNamespace": "UGF.Testing.Runtime",
     "references": [
         "GUID:27619889b8ba8c24980f49ee34dbb44a"
     ],
